+++ src/main/java/de/maxhenkel/easyvillagers/blocks/BreederBlock.java
@@ -5,9 +5,10 @@ import de.maxhenkel.corelib.item.ItemUtils;
 import de.maxhenkel.easyvillagers.Main;
 import de.maxhenkel.easyvillagers.ModItemGroups;
 import de.maxhenkel.easyvillagers.blocks.tileentity.BreederTileentity;
+import de.maxhenkel.easyvillagers.blocks.tileentity.render.BreederRenderer;
 import de.maxhenkel.easyvillagers.gui.BreederContainer;
 import de.maxhenkel.easyvillagers.items.VillagerItem;
-import de.maxhenkel.easyvillagers.items.render.BreederItemRenderer;
+import de.maxhenkel.easyvillagers.items.render.BlockItemRendererBase;
 import net.minecraft.block.BlockRenderType;
 import net.minecraft.block.BlockState;
 import net.minecraft.block.ITileEntityProvider;
@@ -37,13 +38,13 @@ import javax.annotation.Nullable;
 public class BreederBlock extends VillagerBlockBase implements ITileEntityProvider, IItemBlock {
 
     public BreederBlock() {
-        super(Properties.create(Material.IRON).hardnessAndResistance(2.5F).sound(SoundType.METAL).notSolid().setLightLevel(value -> 15));
+        super(Properties.create(Material.IRON).hardnessAndResistance(2.5F).sound(SoundType.METAL).notSolid());
         setRegistryName(new ResourceLocation(Main.MODID, "breeder"));
     }
 
     @Override
     public Item toItem() {
-        return new BlockItem(this, new Item.Properties().group(ModItemGroups.TAB_EASY_VILLAGERS).setISTER(() -> BreederItemRenderer::new)).setRegistryName(getRegistryName());
+        return new BlockItem(this, new Item.Properties().group(ModItemGroups.TAB_EASY_VILLAGERS).setISTER(() -> () -> new BlockItemRendererBase<>(BreederRenderer::new, BreederTileentity::new))).setRegistryName(getRegistryName());
     }
 
     @Override
@@ -114,7 +115,7 @@ public class BreederBlock extends VillagerBlockBase implements ITileEntityProvid
 
     @Override
     public BlockRenderType getRenderType(BlockState state) {
-        return BlockRenderType.MODEL;
+        return BlockRenderType.INVISIBLE;
     }
 
     @OnlyIn(Dist.CLIENT)