+++ src/main/java/de/maxhenkel/easyvillagers/blocks/IncubatorBlock.java
@@ -4,8 +4,9 @@ import de.maxhenkel.corelib.block.IItemBlock;
 import de.maxhenkel.easyvillagers.Main;
 import de.maxhenkel.easyvillagers.ModItemGroups;
 import de.maxhenkel.easyvillagers.blocks.tileentity.IncubatorTileentity;
+import de.maxhenkel.easyvillagers.blocks.tileentity.render.IncubatorRenderer;
 import de.maxhenkel.easyvillagers.gui.VillagerIOContainer;
-import de.maxhenkel.easyvillagers.items.render.IncubatorItemRenderer;
+import de.maxhenkel.easyvillagers.items.render.BlockItemRendererBase;
 import net.minecraft.block.BlockRenderType;
 import net.minecraft.block.BlockState;
 import net.minecraft.block.ITileEntityProvider;
@@ -35,13 +36,13 @@ import javax.annotation.Nullable;
 public class IncubatorBlock extends VillagerBlockBase implements ITileEntityProvider, IItemBlock {
 
     public IncubatorBlock() {
-        super(Properties.create(Material.IRON).hardnessAndResistance(2.5F).sound(SoundType.METAL).notSolid().setLightLevel(value -> 15));
+        super(Properties.create(Material.IRON).hardnessAndResistance(2.5F).sound(SoundType.METAL).notSolid());
         setRegistryName(new ResourceLocation(Main.MODID, "incubator"));
     }
 
     @Override
     public Item toItem() {
-        return new BlockItem(this, new Item.Properties().group(ModItemGroups.TAB_EASY_VILLAGERS).setISTER(() -> IncubatorItemRenderer::new)).setRegistryName(getRegistryName());
+        return new BlockItem(this, new Item.Properties().group(ModItemGroups.TAB_EASY_VILLAGERS).setISTER(() -> () -> new BlockItemRendererBase<>(IncubatorRenderer::new, IncubatorTileentity::new))).setRegistryName(getRegistryName());
     }
 
     @Override
@@ -76,7 +77,7 @@ public class IncubatorBlock extends VillagerBlockBase implements ITileEntityProv
 
     @Override
     public BlockRenderType getRenderType(BlockState state) {
-        return BlockRenderType.MODEL;
+        return BlockRenderType.INVISIBLE;
     }
 
     @OnlyIn(Dist.CLIENT)