+++ src/main/java/de/maxhenkel/easyvillagers/blocks/tileentity/FakeWorldTileentity.java
@@ -1,14 +1,17 @@
 package de.maxhenkel.easyvillagers.blocks.tileentity;
 
+import net.minecraft.block.BlockState;
 import net.minecraft.tileentity.TileEntityType;
 import net.minecraft.world.World;
 
 public class FakeWorldTileentity extends SyncableTileentity {
 
     private boolean fakeWorld;
+    private BlockState defaultState;
 
-    public FakeWorldTileentity(TileEntityType<?> tileEntityTypeIn) {
-        super(tileEntityTypeIn);
+    public FakeWorldTileentity(TileEntityType<?> tileEntityType, BlockState defaultState) {
+        super(tileEntityType);
+        this.defaultState = defaultState;
     }
 
     public void setFakeWorld(World w) {
@@ -18,6 +21,14 @@ public class FakeWorldTileentity extends SyncableTileentity {
 
     public boolean isFakeWorld() {
         return fakeWorld;
+    }
+
+    @Override
+    public BlockState getBlockState() {
+        if (fakeWorld) {
+            return defaultState;
+        }
+        return super.getBlockState();
     }
 
 }