+++ src/main/java/de/maxhenkel/easyvillagers/blocks/tileentity/TraderTileentity.java
@@ -1,10 +1,11 @@
 package de.maxhenkel.easyvillagers.blocks.tileentity;
 
+import de.maxhenkel.easyvillagers.blocks.ModBlocks;
+
 public class TraderTileentity extends TraderTileentityBase {
 
     public TraderTileentity() {
-        super(ModTileEntities.TRADER);
-
+        super(ModTileEntities.TRADER, ModBlocks.TRADER.getDefaultState());
     }
 
 }